--- conda-environment.yml
@@ -1,7 +1,0 @@
-name: env_utilities
-channels:
-  - conda-forge
-  - defaults
-dependencies:
-  - pip==243.1
-  - python=3.11.8
