+++ src/utilities/.tests/validation/test_vgeneral.py
@@ -12,8 +12,8 @@
 from typing import Any, Iterable
 
 # User.
-from exceptions.eiterables import WrongLengthError
-from exceptions.etypes import WrongTypeError
+from src.utilities.exceptions.ecollections import WrongLengthError
+from src.utilities.exceptions.etypes import WrongTypeError
 
 
 # #############################################################################
