+++ src/utilities/general/gstrings.py
