+++ src/utilities/general/gtypes.py
